refactor (student progress page): split the student progress page components into independent modules
Split the components of the student progress page into independent modules, including ProgressTabs, LessonProgress, AssessmentProgress, ProgressMetrics, and ProgressReport, to improve the maintainability and reusability of the code. At the same time, remove the test code in the development environment.
